requirements.txt
pillow
pyscreenshot
pika
schedule

